supplier/requirements.txt
requests
beautifulsoup4
feedparser
newspaper3k
firebase-admin
python-dotenv
lxml_html_clean
flask
aiohttp
playwright
